// packages/openbook/src/types.ts
export interface ArgType {
  name?: string;
  description?: string;
  type?: ArgTypeInfo;
  control?: ControlType;
  options?: unknown[];
  table?: ArgTypeTable;
  if?: ConditionalArg;
}

export interface ArgTypeTable {
  defaultValue?: { summary: string };
  category?: string;
}

export interface ArgTypeInfo {
  name: "string" | "number" | "boolean" | "array" | "object" | "function" | "enum";
  required?: boolean;
}

export interface ControlOptions {
  type: "select" | "radio" | "check";
  options: unknown[];
}

export type ControlType =
  | "boolean"
  | "text"
  | "number"
  | "select"
  | "radio"
  | "check"
  | ControlOptions;

export interface ConditionalArg {
  arg?: string;
  global?: string;
  truthy?: boolean;
  exists?: boolean;
  eq?: unknown;
  neq?: unknown;
}